added work-flow

+++ mace4/msearch.h
@@ -130,7 +130,7 @@ private:
   std::string interp_file_name;
   ofstream* models_interp_file_stream = nullptr;
   size_t out_models_count = 0;
-  size_t max_count = 0;
+  int    max_count = 0;
   size_t file_count = 1;
   int  id2val(int id);
   int  f0_val(int base);
